Complete bounded package occurrence scanning
